refactor the code and introduce SafeProcessHandle.Start: Unix part
