+++ src/libraries/System.Diagnostics.Process/src/Microsoft/Win32/SafeHandles/SafeProcessHandle.Unix.cs
@@ -1,17 +1,16 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-/*============================================================
-**
-** Class:  SafeProcessHandle
-**
-** A wrapper for a process handle
-**
-**
-===========================================================*/
-
-using System;
-using System.Diagnostics;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.IO;
+using System.IO.Pipes;
+using System.Runtime.InteropServices;
+using System.Runtime.Versioning;
+using System.Security;
+using System.Text;
+using System.Threading;
+using Microsoft.Win32.SafeHandles;
 
 namespace Microsoft.Win32.SafeHandles
 {
@@ -45,14 +44,166 @@ namespace Microsoft.Win32.SafeHandles
             return true;
         }
 
-        private static SafeProcessHandle StartCore(ProcessStartInfo startInfo, SafeFileHandle? stdinHandle, SafeFileHandle? stdoutHandle, SafeFileHandle? stderrHandle)
+        internal static SafeProcessHandle StartCore(ProcessStartInfo startInfo, SafeFileHandle? stdinHandle, SafeFileHandle? stdoutHandle, SafeFileHandle? stderrHandle)
         {
-            _ = startInfo;
-            _ = stdinHandle;
-            _ = stdoutHandle;
-            _ = stderrHandle;
+            if (ProcessUtils.PlatformDoesNotSupportProcessStartAndKill)
+            {
+                throw new PlatformNotSupportedException();
+            }
 
-            throw new NotImplementedException();
+            ProcessUtils.EnsureInitialized();
+
+            string? filename;
+            string[] argv;
+
+            string[] envp = ProcessUtils.CreateEnvp(startInfo);
+            string? cwd = !string.IsNullOrWhiteSpace(startInfo.WorkingDirectory) ? startInfo.WorkingDirectory : null;
+
+            bool setCredentials = !string.IsNullOrEmpty(startInfo.UserName);
+            uint userId = 0;
+            uint groupId = 0;
+            uint[]? groups = null;
+            if (setCredentials)
+            {
+                (userId, groupId, groups) = ProcessUtils.GetUserAndGroupIds(startInfo);
+            }
+
+            // .NET applications don't echo characters unless there is a Console.Read operation.
+            // Unix applications expect the terminal to be in an echoing state by default.
+            // To support processes that interact with the terminal (e.g. 'vi'), we need to configure the
+            // terminal to echo. We keep this configuration as long as there are children possibly using the terminal.
+            // Handle can be null only for UseShellExecute or platforms that don't support Console.Open* methods like Android.
+            bool usesTerminal = (stdinHandle is not null && Interop.Sys.IsATty(stdinHandle))
+                || (stdoutHandle is not null && Interop.Sys.IsATty(stdoutHandle))
+                || (stderrHandle is not null && Interop.Sys.IsATty(stderrHandle));
+
+            if (startInfo.UseShellExecute)
+            {
+                string verb = startInfo.Verb;
+                if (verb != string.Empty &&
+                    !string.Equals(verb, "open", StringComparison.OrdinalIgnoreCase))
+                {
+                    throw new Win32Exception(Interop.Errors.ERROR_NO_ASSOCIATION);
+                }
+
+                // On Windows, UseShellExecute of executables and scripts causes those files to be executed.
+                // To achieve this on Unix, we check if the file is executable (x-bit).
+                // Some files may have the x-bit set even when they are not executable. This happens for example
+                // when a Windows filesystem is mounted on Linux. To handle that, treat it as a regular file
+                // when exec returns ENOEXEC (file format cannot be executed).
+                filename = ProcessUtils.ResolveExecutableForShellExecute(startInfo.FileName, cwd);
+                if (filename != null)
+                {
+                    argv = ProcessUtils.ParseArgv(startInfo);
+
+                    SafeProcessHandle processHandle = ForkAndExecProcess(
+                        startInfo, filename, argv, envp, cwd,
+                        setCredentials, userId, groupId, groups,
+                        stdinHandle, stdoutHandle, stderrHandle, usesTerminal,
+                        throwOnNoExec: false); // return invalid handle instead of throwing on ENOEXEC
+
+                    if (!processHandle.IsInvalid)
+                    {
+                        return processHandle;
+                    }
+                }
+
+                // use default program to open file/url
+                filename = GetPathToOpenFile();
+                argv = ProcessUtils.ParseArgv(startInfo, filename, ignoreArguments: true);
+
+                return ForkAndExecProcess(
+                    startInfo, filename, argv, envp, cwd,
+                    setCredentials, userId, groupId, groups,
+                    stdinHandle, stdoutHandle, stderrHandle, usesTerminal);
+            }
+            else
+            {
+                filename = ProcessUtils.ResolvePath(startInfo.FileName);
+                argv = ProcessUtils.ParseArgv(startInfo);
+                if (Directory.Exists(filename))
+                {
+                    throw new Win32Exception(SR.DirectoryNotValidAsInput);
+                }
+
+                return ForkAndExecProcess(
+                    startInfo, filename, argv, envp, cwd,
+                    setCredentials, userId, groupId, groups,
+                    stdinHandle, stdoutHandle, stderrHandle, usesTerminal);
+            }
+        }
+
+        private static SafeProcessHandle ForkAndExecProcess(
+            ProcessStartInfo startInfo, string? resolvedFilename, string[] argv,
+            string[] envp, string? cwd, bool setCredentials, uint userId,
+            uint groupId, uint[]? groups,
+            SafeFileHandle? stdinHandle, SafeFileHandle? stdoutHandle, SafeFileHandle? stderrHandle,
+            bool usesTerminal, bool throwOnNoExec = true)
+        {
+            if (string.IsNullOrEmpty(resolvedFilename))
+            {
+                Interop.ErrorInfo errno = Interop.Error.ENOENT.Info();
+                throw ProcessUtils.CreateExceptionForErrorStartingProcess(errno.GetErrorMessage(), errno.RawErrno, startInfo.FileName, cwd);
+            }
+
+            // Lock to avoid races with OnSigChild
+            // By using a ReaderWriterLock we allow multiple processes to start concurrently.
+            ProcessUtils.s_processStartLock.EnterReadLock();
+            try
+            {
+                if (usesTerminal)
+                {
+                    ProcessUtils.ConfigureTerminalForChildProcesses(1);
+                }
+
+                int childPid;
+
+                // Invoke the shim fork/execve routine.  It will fork a child process,
+                // map the provided file handles onto the appropriate stdin/stdout/stderr
+                // descriptors, and execve to execute the requested process.  The shim implementation
+                // is used to fork/execve as executing managed code in a forked process is not safe (only
+                // the calling thread will transfer, thread IDs aren't stable across the fork, etc.)
+                int errno = Interop.Sys.ForkAndExecProcess(
+                    resolvedFilename, argv, envp, cwd,
+                    setCredentials, userId, groupId, groups,
+                    out childPid, stdinHandle, stdoutHandle, stderrHandle);
+
+                if (errno == 0)
+                {
+                    // Ensure we'll reap this process.
+                    // note: SetProcessId will set this if we don't set it first.
+                    _waitStateHolder = new ProcessWaitState.Holder(childPid, isNewChild: true, usesTerminal);
+
+                    // Store the child's information into this Process object.
+                    Debug.Assert(childPid >= 0);
+                    SetProcessId(childPid);
+                    SetProcessHandle(new SafeProcessHandle(_processId, GetSafeWaitHandle()));
+
+                    return true;
+                }
+                else
+                {
+                    if (!throwOnNoExec &&
+                        new Interop.ErrorInfo(errno).Error == Interop.Error.ENOEXEC)
+                    {
+                        return false;
+                    }
+
+                    throw ProcessUtils.CreateExceptionForErrorStartingProcess(new Interop.ErrorInfo(errno).GetErrorMessage(), errno, resolvedFilename, cwd);
+                }
+            }
+            finally
+            {
+                ProcessUtils.s_processStartLock.ExitReadLock();
+
+                if (_waitStateHolder == null && usesTerminal)
+                {
+                    // We failed to launch a child that could use the terminal.
+                    ProcessUtils.s_processStartLock.EnterWriteLock();
+                    ProcessUtils.ConfigureTerminalForChildProcesses(-1);
+                    ProcessUtils.s_processStartLock.ExitWriteLock();
+                }
+            }
         }
     }
 }
