+++ src/libraries/System.Diagnostics.Process/src/System/Diagnostics/ProcessUtils.ConfigureTerminalForChildProcesses.iOS.cs
@@ -3,7 +3,7 @@
 
 namespace System.Diagnostics
 {
-    public partial class Process
+    internal static partial class ProcessUtils
     {
         /// These methods are used on other Unix systems to track how many children use the terminal,
         /// and update the terminal configuration when necessary.
